+++ dsr_agents/CHANGELOG.rst
@@ -13,6 +13,8 @@ X.X.X (XX-02-2024)
 * Refactor CMakelists.txt.
 * Added new functions to agent node.
 * Remove viewer as it's integrated in dsr_rqt_plugin.
+* Remove QtExecutor as it's integrated in dsr_util.
+* Move ros_to_dsr_types.hpp to dsr_util.
 
 0.6.2 (23-01-2024)
 ------------------